requirements.txt
httpx==0.27.0
python-telegram-bot==21.5
apscheduler==3.10.4
